[2.3.x] Remove redundant cast. Remove manual array and collection copies.

+++ modules/http-ajp/src/test/java/org/glassfish/grizzly/http/ajp/BasicAjpTest.java
@@ -535,7 +535,7 @@ public class BasicAjpTest extends AjpTestBase {
                 
                 if (isOk) {
                     try {
-                        assertEquals((Integer) 256, request.getAttribute(SSLSupport.KEY_SIZE_KEY));
+                        assertEquals(256, request.getAttribute(SSLSupport.KEY_SIZE_KEY));
                         assertNotNull(request.getAttribute(SSLSupport.SESSION_ID_KEY));
                         assertNotNull(request.getAttribute(SSLSupport.CIPHER_SUITE_KEY));
                         assertNotNull(request.getAttribute(SSLSupport.CERTIFICATE_KEY));
